+++ apps/antalmanac/src/components/Header/ProfileMenuButtons.tsx
@@ -1,11 +1,11 @@
-import { FC } from 'react';
-
+import { AccountCircle, Menu } from '@mui/icons-material';
 import { Button, IconButton, CircularProgress } from '@mui/material';
-
-import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import Image from 'next/image';
-import MenuIcon from '@mui/icons-material/Menu';
 import { User } from '@packages/antalmanac-types';
+import Image from 'next/image';
+import { useEffect, useState } from 'react';
+
+import AppStore from '$stores/AppStore';
+
 
 interface ProfileMenuButtonsProps {
     user: User | null;
@@ -13,21 +13,37 @@ interface ProfileMenuButtonsProps {
     handleSettingsOpen: (event: React.MouseEvent<HTMLElement>) => void;
     loading?: boolean;
 }
-export const ProfileMenuButtons: FC<ProfileMenuButtonsProps> = ({ user, handleOpen, handleSettingsOpen, loading = false }) => {
+
+export function ProfileMenuButtons({ user, handleOpen, handleSettingsOpen, loading = false }: ProfileMenuButtonsProps) {
+    const [skeletonMode, setSkeletonMode] = useState(AppStore.getSkeletonMode());
+
+    useEffect(() => {
+        const handleSkeletonModeChange = () => {
+            setSkeletonMode(AppStore.getSkeletonMode());
+        };
+
+        AppStore.on('skeletonModeChange', handleSkeletonModeChange);
+
+        return () => {
+            AppStore.off('skeletonModeChange', handleSkeletonModeChange);
+        };
+    }, []);
+
     if (!user) {
         return (
             <>
                 <Button
                     variant="text"
                     size="medium"
-                    startIcon={loading ? <CircularProgress size={20} color="inherit" /> : <AccountCircleIcon />}
+                    startIcon={loading ? <CircularProgress size={20} color="inherit" /> : <AccountCircle />}
                     color="inherit"
                     onClick={handleOpen}
+                    disabled={skeletonMode}
                 >
                     Sign In
                 </Button>
                 <IconButton onClick={handleSettingsOpen} color="inherit">
-                    <MenuIcon />
+                    <Menu />
                 </IconButton>
             </>
         );
@@ -35,31 +51,33 @@ export const ProfileMenuButtons: FC<ProfileMenuButtonsProps> = ({ user, handleOp
 
     const { name, avatar } = user;
 
-    const profileButtonStyles = {
-        display: 'flex',
-        alignItems: 'center',
-        gap: '4px',
-        padding: '6px 8px',
-        borderRadius: 24,
-        border: 'none',
-    };
-
-    const profilePicStyles = {
-        width: 24,
-        height: 24,
-        borderRadius: '100%',
-        color: 'inherit',
-        whiteSpace: 'normal',
-    };
-
     return (
-        <Button sx={profileButtonStyles} onClick={handleOpen} variant="text" color="inherit">
+        <Button
+            sx={{
+                display: 'flex',
+                alignItems: 'center',
+                gap: '4px',
+                padding: '6px 8px',
+                borderRadius: 24,
+                border: 'none',
+            }}
+            onClick={handleOpen}
+            variant="text"
+            color="inherit"
+            disabled={skeletonMode}
+        >
             {avatar ? (
-                <Image src={avatar} alt={name ?? 'User avatar'} width={24} height={24} style={profilePicStyles} />
+                <Image
+                    src={avatar}
+                    alt={name ?? 'User avatar'}
+                    width={24}
+                    height={24}
+                    style={{ width: 24, height: 24, borderRadius: '100%', color: 'inherit', whiteSpace: 'normal' }}
+                />
             ) : (
-                <AccountCircleIcon sx={{ width: 24, height: 24 }} />
+                <AccountCircle sx={{ width: 24, height: 24 }} />
             )}
-            <MenuIcon />
+            <Menu />
         </Button>
     );
-};
+}
